Baseline the format_text() FORMAT_HTML PHPDoc mismatch instead of casting
format_text()'s docblock declares $format as int, but Moodle core defines
FORMAT_HTML as the string '1' (lib/weblib.php), so passing the constant
directly trips PHPStan. This project already has a convention for this
exact class of pre-existing core PHPDoc inaccuracy: baseline it, scoped by
path, rather than work around it in our code (see the existing
moodleform_mod::standard_intro_elements() entry, documented in
CHANGELOG.md). Revert the local (int) cast at save_entry.php:106 and add a
matching phpstan-baseline.neon entry instead.

+++ classes/external/save_entry.php
@@ -103,7 +103,7 @@ class save_entry extends external_api {
             'id' => $id,
             'timemodified' => $now,
             'timestr' => $timestr,
-            'responsehtml' => format_text($response, (int) FORMAT_HTML, ['context' => $context]),
+            'responsehtml' => format_text($response, FORMAT_HTML, ['context' => $context]),
         ];
     }
